+++ vandal/__main__.py
@@ -1,15 +1,19 @@
-# imports all CLI clients.
+# imports all relevant contents.
 import vandal
-from vandal.objects import montecarlo 
+from vandal.objects import montecarlo
+from colorama import Fore
+import colorama
 
-# runs the main CLI client.
+colorama.init()
+
+# runs the main CLI client that leads to individual apps.
 if __name__ == '__main__':
     import argparse
     parser = argparse.ArgumentParser()
-    parser.add_argument('-o', '--option', help = 'Enter options', type = str, 
+    parser.add_argument('-e', '--entry', help = 'Enter options', type = str,
     choices = ['montecarlo'])
     args = parser.parse_args()
 
-    if args.option == 'montecarlo':
-        vandal.objects.montecarlo.Main()
-
+    if args.entry == 'montecarlo':
+        print(Fore.GREEN + '=== ENTERING MONTECARLO CLIENT... ===\n', Fore.RESET)
+        vandal.objects.montecarlo.MCapp()