+++ vandal/objects/__init__.py
@@ -1,5 +1,8 @@
 # import all relevant contents from the associated module.
-from vandal.objects.montecarlo import MonteCarlo
+from vandal.objects.montecarlo import (
+    MonteCarlo,
+    MCapp,
+)
 from vandal.objects.eoq import EOQ
 from vandal.objects.dijkstra import Dijkstra
 
@@ -8,4 +11,5 @@ __all__ = [
     MonteCarlo,
     EOQ,
     Dijkstra,
+    MCapp,
 ]